pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "medtriage"
version = "1.0.0"
description = "MedTriage: OpenEnv Clinical ED Triage Benchmark - Emergency Department Triage Agent Evaluation"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "MedTriage Team", email = "info@medtriage.example.com"}
]
keywords = ["openenv", "benchmark", "clinical", "triage", "emergency-department", "healthcare", "llm"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Healthcare Industry",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Medical/Dental",
]
dependencies = [
    "pydantic>=2.0",
    "openai>=1.0.0",
    "scipy>=1.10.0",
    "numpy>=1.24.0",
    "pytest>=7.0.0",
    "rich>=13.0.0",
    "typer>=0.9.0",
    "python-dotenv>=1.0.0",
    "flask>=3.0.0",
    "openenv-core>=0.2.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0",
    "flake8>=6.0",
    "mypy>=1.0",
    "pytest-cov>=4.0",
]

[project.urls]
repository = "https://github.com/Santhosh121805/meditrage"
documentation = "https://github.com/Santhosh121805/meditrage/blob/main/README.md"
homepage = "https://huggingface.co/spaces/sandy0042v/meditage"

[project.scripts]
medtriage = "inference:main"
server = "server.app:run_server"

[tool.setuptools]
packages = ["models", "graders", "tasks", "utils", "tests"]

[tool.setuptools.package-data]
medtriage = ["data/*.json", "*.yaml"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-v --tb=short"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
